Move Contraints to Constraint.
