+++ src/Querys/QueryBuilder.php
@@ -9,8 +9,8 @@ use Yiisoft\Db\Drivers\Connection;
 use Yiisoft\Db\Exceptions\Exception;
 use Yiisoft\Db\Exceptions\InvalidConfigException;
 use Yiisoft\Db\Querys\Conditions\ConditionInterface;
-use Yiisoft\Db\Constraints\Constraint;
-use Yiisoft\Db\Constraints\ConstraintFinderInterface;
+use Yiisoft\Db\Constraint\Constraint;
+use Yiisoft\Db\Constraint\ConstraintFinderInterface;
 use Yiisoft\Db\Exceptions\InvalidArgumentException;
 use Yiisoft\Db\Exceptions\NotSupportedException;
 use Yiisoft\Db\Expressions\Expression;
